requirements.txt
SQLAlchemy-Utils[babel, url] >= 0.36.8
fastapi[all]
sqlalchemy-stubs